Collusion with wall,  crosshair & audio
fixed wall collusion and there is now audio on gun and empty ammo, and there is a crosshair
