+++ Assets/_Data/WeaponData/BulletData/AssultRifleWeaponData.asset
@@ -15,7 +15,7 @@ MonoBehaviour:
   <BulletData>k__BackingField: {fileID: 11400000, guid: 96db7f8127f708544b5420209ba6642e, type: 2}
   <AmmoCapacity>k__BackingField: 100
   <AutomaticFire>k__BackingField: 1
-  <WeaponDelay>k__BackingField: 0.2
+  <WeaponDelay>k__BackingField: 0.1
   <SpreadAngle>k__BackingField: 3
   multiBulletShoot: 0
   bulletCount: 1
