+++ Assets/_Data/WeaponData/BulletData/RegularBullet.asset
@@ -21,5 +21,5 @@ MonoBehaviour:
   <IsRayCast>k__BackingField: 0
   <ImpactObstaclePrefab>k__BackingField: {fileID: 0}
   <ImpactEnemyPrefab>k__BackingField: {fileID: 0}
-  <KnockBackPower>k__BackingField: 6
+  <KnockBackPower>k__BackingField: 2
   <KnockBackDelay>k__BackingField: 0.1
