+++ Assets/_scripts/AudioPlayer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ecaa3d2a175596f44a32abcc883b4a52
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
