+++ Assets/_scripts/Weapons/WeaponAudio.cs
@@ -1,0 +1,20 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class WeaponAudio : AudioPlayer
+{
+    [SerializeField]
+    private AudioClip shootBulletClip = null, outOfBulletsClip = null;
+
+    public void PlayShootSound()
+    {
+        PlayClip(shootBulletClip);
+    }
+
+    public void PlayNoBulletSound()
+    {
+        PlayClip(outOfBulletsClip);
+    }
+
+}
