+++ Assets/_scripts/Weapons/WeaponAudio.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 25955377ce8608f4799adfdc6c644ccd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
